src/util/util.lisp: dump-defsystem-forms: tweak lispworks personal
* src/util/util.lisp: (if-feature, if-not-feature): new
defstructs. (print-feature-hack): struct object printer for if-feature
and if-not-feature structs. this is ahack to insert reader
conditionals in lisp forms. (make-bindings-module): make the
bindings-file :load-only on lispworks-personal edition.

This will eventually end up in the make-defsystems forms.

The downside is if you are on lispworks-personal, then one can no
longer call CLAW.UTIL:CLAW-CXX-DEFSYSTEMS to define the defsystem
through the call to mk:defsystem programatically anymore as that will
fail.  Instead one has to call (macroexpand '(claw-cxx:defsystem ..))
and then dump the $claw-cxx-defsystems-source-form to a system
definition file and load that instead.